+++ terra_geocrud/serializers.py
@@ -288,10 +288,10 @@ class CrudFeatureDetailSerializer(BaseUpdatableMixin, FeatureSerializer):
         routing_information = validated_data.pop('routing_information', {})
 
         feature = Feature.objects.create(**validated_data)
-        models.RoutingInformations.objects.get_or_create(feature=feature,
-                                                         defaults={'route_description': routing_information.get(
-                                                             'route_description',
-                                                             {})})
+        models.RoutingInformations.objects.create(feature=feature,
+                                                  defaults={'route_description': routing_information.get(
+                                                      'route_description',
+                                                      {})})
         return feature
 
     def update(self, instance, validated_data):